refactor(cli): rename cli package to cmd
