+++ pkg/cmd/run.go
@@ -1,4 +1,4 @@
-package cli
+package cmd
 
 import (
 	"errors"
